U-boot NAND Add 512 byte SW ECC support
This is compatible with the android kernel.

These are needed for flashing of android filesystem using fastboot so that
kernel recognizes the data flashed from fastboot.

In order to use the 512 byte ecc, the board config file sets
This is similar to how xloader is configured.

If CFG_SW_ECC_512 is not set, the older 256 byte sw ecc is used.

The 512 ecc algorithm is in the nand_ecc_512.c file.
The 256 ecc algorithm is in the nand_ecc_256.c file.

Once all boards of interest are using the newer ecc, these
files can be reduced to just nand_ecc.c

This was run tested on Zoom1 with the 512 byte ECC xloader
This was compile tested on Zoom2 and SPD.

Signed-off-by: Vimal Singh <vimalsingh@ti.com>
Signed-off-by: Tom Rix <Tom.Rix@windriver.com>
Ack-by: Nishanth Menon <nm@ti.com>