U-boot:FASTBOOT Add support for write.i and write.yaffs
The kernel partition is now writen with write.i
The system, user data and cache partitions are now written with write.yaffs

From the fastboot client the kernel is flashed with

./fastboot_client flash kernel <path to kernel>

The system yaffs2 image is flashed with

./fastboot_client flash system <path to yaffs2 image>

Fastboot takes care of where and how to flash.
The offset and the size of the image is stored as an environment variable

For the kernel, the variable are
kernel_nand_offset
kernel_nand_size

Similar for the system.

These can be used later such as loading the kernel

nand read.i 0x80000000 ${kernel_nand_offset} ${kernel_nand_size}; bootm

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>