U-boot fastboot change bulk endpoint size at runtime.
Use the MUSB_POWER_HSMODE bit to configure the bulk end buffer size.  When
the bit is set (high speed), use a size of 512 bytes for the receive buffer.
When the bit is clear (full speed), use the size of 64 bytes for the receive
buffer.

This allows the fastboot to work with usb 1.1 hubs without expliciting setting
CONFIG_USB_1_1_DEVICE.

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>