U-boot Fastboot Change NAND partition sizes.
The environment partition offset changed to 0x1c0000.  This overlaps the
existing kernel parition.  Change the kernel parition be a the end of the
new environment partition.  Recent android multi images exceed the exiting
kernel partition size of 4M so increase the kernel size to 30M. Increase the
system partition, 128M to 160M, by a little to use all the unused NAND.

The kernel parition values should now be

0x00000000-0x00080000 : "X-Loader-NAND"
0x00080000-0x00200000 : "U-Boot-NAND"
0x001c0000-0x00200000 : "Boot Env-NAND"
0x00200000-0x01e00000 : "Kernel-NAND"
0x02000000-0x0c000000 : "system"
0x0c000000-0x0e000000 : "userdata"
0x0e000000-0x10000000 : "cache"

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>