U-boot OMAP3: Add support for OMAP3 die ID
Read and store OMAP3 die ID in U-Boot environment.

This patch is cherry picked from the mainline u-boot

commit e6a6a704151c2d7e4a7b485545b48a6020ccca17
Author: Dirk Behme <dirk.behme@googlemail.com>
Date:   Thu Mar 12 19:30:50 2009 +0100

This mainline change has a better interface to device identifier registers
than what is currently being done.

The original had a printout of the dieid as part of the startup banner.
This print out is a distraction and was removed

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>