U-boot FASTBOOT Add runtime partitions
The fastboot partition table can be added to at runtime.
The evironment variable 'fbparts' is queried when fastboot
is initialized.  Partitions specified in the fbparts will
be used during the fastboot session.

The format of 'fbparts' is similar to the 'mtdparts' used
to set linux kernel mtd partitions from the bootargs.

The parsing of the fbparts is derrived from the 2.6.27
linux kernel drivers/mtd/cmdlinepart.c newpart function

An example of its use is

fbparts 512k@0(xloader)hwecc,\
	536k@512k(bootloader)swecc,\
	29m@2m(kernel)i|swecc,\
	160m@32m(system)yaffs|swecc,\
	32m@192m(userdata)yaffs|swecc,\
	32m@224m(cache)yaffs|swecc

The corresponding mtdparts added to the bootargs is

mtdparts=omap2-nand.0:512k@0(xloader),\
	1536k@512k(bootloader),\
	29m@2m(kernel),\
	160m@32m(system),\
	32m@192m(userdata),\
	32m@224m(cache)

The format is
	<size>@<offset>(<name>)<optional-flags>,

The <name> is the same as what is used by the fastboot client.

The supported flags are
	i	: use nand write.i
	yaffs	: use nand write.yaffs
	swecc	: use nand swecc, software ecc
	hwecc	: use nand hwecc, hardware ecc

Multiple flags can be or-ed together with a '|'

A runtime partition will not be used if it overlaps a static
partition.

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>