Fastboot support disconnecting and reconnecting the usb cable
Earlier, when the cable was disconnected, the fastboot device
shutdown.  Now that the user can disconnect either from a
timeout or a Ctrl-C on the console, do not shutdown the device
with a disconnect is detected. Go back to the start usb
discovery state and wait for the cable to be reconnected.

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>