Fastboot Fix warning about invalid buffer size.
In linux syslog, there is a warning that the existing size of 64
bytes for the tx fifo is incorrect.  The fastboot client only reads
at most 64 bytes so this not a true error.  However to quiet the
warning make the tx fifo size behave the same as the rx fifo.
Both now are 512 bytes for high speed and and 64 bytes for fullspeed.

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>