U-boot OMAP3: Add support for USB, AC charger detection
This patch will check for the presence of a USB charger or
an AC T2 charger. If a charger is detected the blue led blinks
In the absence of a AC T2 or a USB charger, an assumption is
made that the board is powered by battery. If the battery charge
is <3.3V, the red led blinks and the kernel is prevented from booting.
If AC debug board charger is connected, the battery voltage is
approximately 4.1V and the kernel can boot.

Signed-off-by: Christina Warren <cawarren@ti.com>