U-boot OMAP3: Modify the default bootargs and bootcmd
This patch will modify the default bootcmd to check for the
presence of an mmc card. If present, we will boot out of mmc
else we will boot out of nand. This change is only for Zoom2.

Note: These arguments will be used only when the user does not
have any boot args already specified.

Signed-off-by: Christina Warren <cawarren@ti.com>