OMAP36XX Fix the extents of PER clock parameters
Some of the bit extents of the PER clock parameters increased by 1 bit
over 34xx.  Add #defines for these bit sizes.  Use the defines in
the 36xx PER clk init and clock info command.

They are defined but not used for 34xx to make the review of 34xx code
easier.

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>