Bug-fix: Cannot disable I2C module inbetween of a transaction. Fixes the I2C read errors at L3=200MHz.
This patch fixes a bug that was causing problems for I2C operation
at L3=200MHz.

I2C read is a two phase operation:
1. Master writes the register offset to the Slave
2. Slave provides the data from specified register to Master

These two operations happen as a single transaction (stop condition
generated only after the final read) and the I2C module cannot be
disabled or the bus cannot be free'd inbetween. This was the bug
that is fixed now by this patch.

Disabling the module inbetween leaves the I2C controller in an
unpredictable state and thereby transfer might fail resulting
in errors.

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>
Signed-off-by: Tom Rix <Tom.Rix@windriver.com>