u-boot: mmcinit reboot fix
On reboot from the kernel, mmc does not get initialized in u-boot.
RESET DONE bit is not set because fck and ick are not running on system
reboot. According to TRM, RESET DONE requires fck, ick and debounce
clock
need to run. Therefore this patch sets fck and ick for all 3 MMC
interfaces.

Signed-off-by: Subramani Venkatesh <subramani.venkatesh@windriver.com>