Fastboot: Fix MUSB DMA word alignment issue on OMAP3630Zoom3
With dma enabled, we were losing the first two bytes of each
page. Due to which the ecc did not match up and the ROM code
was unable to boot the x-loader out of NAND.

This problem was caused by a misaligned address being passed
to the DMA_ADDR register. The MUSB spec states that the start
address (written to the MUSB_DMA_ADDR_CH(n) register) must be
word aligned. This patch fixes the problem.

Signed-off-by: Christina Warren <cawarren@ti.com>