Optimizing NAND read/write to reduce boot time
NAND read accesses are done by reading a single GPMC
register. Address of this register is stored in a
structure and is not changed when reading or writing a
block of data. With the help of this patch, the
register address will now be read before the loop and
not inside the loop.

Signed-off-by: Cyril Bianconi <c-bianconi@ti.com>
Signed-off-by: Sukumar Ghorai<s-ghorai@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>