Fastboot: Add support for new 'upload' command.
The 'upload' function is similar to 'download'.
The direction of data transfer is reversed where the
client uses 'download' to send parition data to the
device, the client can now use 'upload' to receive
partition data from the device. Example: Making a copy
of the kernel partition. The protocol is:
upload:partition_name:buffer_size

The client is expected to query the device for the
correct buffer size by issuing the command:
upload:partition_name:0
The device will return a response DATA<buffer_size>,
similar to what is done for download. If the
buffer_size matches, the device will resend
DATA<buffer_size> and a stream of data that can be
read in similar manner as the status responses, as
received data from the bulk endpoint.

There is a variation of read called 'uploadraw'. The
partition data and the oob data can be read using this
command. It behaves the same as 'upload', except that
the oob data is packed within the regular nand data and
is suitable for use in nand write.raw buff.raw

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>